Start using fixed width integer types

+++ editor/src/main_menu_bar.cpp
@@ -19,7 +19,7 @@ namespace editor
 			{
 				if (ImGui::MenuItem("Open project"))
 				{
-					constexpr unsigned int max_filename_len = 1024;
+					constexpr u16 max_filename_len = 1024;
 					char filename[max_filename_len];
 					FILE* file = popen("zenity --file-selection", "r");
 					fgets(filename, max_filename_len, file);
